requirements.txt
django==4.1.6 

djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0


PyJWT==2.6.0

social-auth-app-django==4.0.0
social-auth-core==4.3.0

dj-rest-auth==2.2.5

django-allauth==0.51.0

python-dotenv==0.21.0

drf-yasg==1.21.4

django-safedelete==1.3.1

pymysql==1.0.2

Pillow==9.3

pandas==1.5.3

openpyxl==3.1.1

